update: contour 配准代码

+++ data_process/data_utils.py
@@ -1,13 +1,16 @@
 import argparse
+import gc
 import os
 import random
 import shutil
+import traceback
 from concurrent.futures import ThreadPoolExecutor, as_completed
 
 import cv2
 import numpy as np
 from PIL import Image
 from loguru import logger
+from matplotlib import pyplot as plt
 from skimage.metrics import structural_similarity as ssim
 from sklearn.model_selection import train_test_split
 
@@ -87,25 +90,31 @@ def ann_filter():
             logger.info(f'{he}没有腺体')
 
 
+he_dir = rf'/NAS2//Data1/lbliao/Data/MSI/pair/1024/M1/tmp/he'
+ihc_dir = rf'/NAS2/Data1/lbliao/Data/MSI/pair/1024/M1/tmp/ihc'
+
+
 # 图像大小过滤
 def size_filter(size: int):
-    ann_dir = r'/data2/lbliao/Data/前列腺癌数据/CKPan/pair/1024-1/tmp/label/'
-
-    # 遍历目录
-    for filename in os.listdir(he_dir):
+    logger.info('start P2')
+    he_dir = rf'/NAS2/Data1/lbliao/Data/MSI/0327/pair/4096/P2/trainA'
+    ihc_dir = rf'/NAS2/Data1/lbliao/Data/MSI/0327/pair/4096/P2/trainB'
+    # ann_dir = r'/data2/lbliao/Data/前列腺癌数据/CKPan/pair/1024-1/tmp/label/'
+    files = os.listdir(he_dir)
+    for filename in files:
         base, _ = os.path.splitext(filename)
         he_path = os.path.join(he_dir, filename)
         ihc_path = os.path.join(ihc_dir, filename)
-        ann_path = os.path.join(ann_dir, f"{base}.txt")
+        # ann_path = os.path.join(ann_dir, f"{base}.txt")
 
         if os.path.isfile(he_path) and os.path.isfile(ihc_path):
             filesize = os.path.getsize(ihc_path)
 
-            if filesize < size * 1024:
+            if filesize < size * 1024 and random.random() < 1:
                 os.remove(he_path)
                 os.remove(ihc_path)
-                if os.path.isfile(ann_path):
-                    os.remove(ann_path)
+                # if os.path.isfile(ann_path):
+                #     os.remove(ann_path)
                 logger.info(f"Deleted {filename} due to size less than {size}KB")
 
 
@@ -177,8 +186,10 @@ def gland_filter():
 
 
 def miss_filter():
-    train_a_dir = he_dir
-    train_b_dir = ihc_dir
+    # train_a_dir = he_dir
+    # train_b_dir = ihc_dir
+    train_a_dir = '/data2/lbliao/Data/MSI/pair/4096/M6/trainA'
+    train_b_dir = '/data2/lbliao/Data/MSI/pair/4096/M6/trainB'
     he_images = os.listdir(train_a_dir)
     ihc_images = os.listdir(train_b_dir)
     for img in he_images:
@@ -195,19 +206,22 @@ def miss_filter():
 
 # parallel_run()
 # gland_filter()
-miss_filter()
+# miss_filter()
 
 
-def split_data():
-    train_a_dir = os.path.join(pair_dir, f'{ihc_ext}/trainA')
-    train_b_dir = os.path.join(pair_dir, f'{ihc_ext}/trainB')
-    test_a_dir = os.path.join(pair_dir, f'{ihc_ext}/testA')
-    test_b_dir = os.path.join(pair_dir, f'{ihc_ext}/testB')
+def split_data_a():
+    he_dir = rf'/NAS2/Data1/lbliao/Data/MSI/0327/pair/4096/P2/he'
+    ihc_dir = rf'/NAS2/Data1/lbliao/Data/MSI/0327/pair/4096/P2/ihc'
+    pair_dir = rf'/NAS2/Data1/lbliao/Data/MSI/0327/dataset/4096/P2'
+    train_a_dir = os.path.join(pair_dir, f'trainA')
+    train_b_dir = os.path.join(pair_dir, f'trainB')
+    test_a_dir = os.path.join(pair_dir, f'testA')
+    test_b_dir = os.path.join(pair_dir, f'testB')
     for directory in [train_a_dir, train_b_dir, test_a_dir, test_b_dir]:
         os.makedirs(directory, exist_ok=True)
 
     images = os.listdir(he_dir)
-    train_set, test_set = train_test_split(images, test_size=0.2, random_state=42)
+    train_set, test_set = train_test_split(images, test_size=0.1, random_state=42)
     for img in train_set:
         if os.path.isfile(os.path.join(he_dir, img)) and os.path.isfile(os.path.join(ihc_dir, img)):
             shutil.copy(os.path.join(he_dir, img), os.path.join(train_a_dir, img))
@@ -219,4 +233,160 @@ def split_data():
             shutil.copy(os.path.join(ihc_dir, img), os.path.join(test_b_dir, img))
 
 
-split_data()
+# split_data()
+# size_filter(8 * 1024)
+
+he_dir = rf'/NAS2/Data1/lbliao/Data/MSI/pair/4096/M6/trainA'
+ihc_dir = rf'/NAS2/Data1/lbliao/Data/MSI/pair/4096/M6/trainB'
+
+
+def is_solid_color_numpy(image_path):
+    # 打开图片并转换为NumPy数组
+    with Image.open(image_path) as img:
+        img_array = np.array(img)
+
+    # 计算像素值的方差
+    variance = np.var(img_array)
+
+    # 如果方差为0，说明所有像素值相同，图片是纯色的
+    return variance == 0
+
+
+def manual_check():
+    images = os.listdir(he_dir)
+    for i, img in enumerate(images):
+        if i < 400:
+            continue
+        try:
+            he_path = os.path.join(he_dir, img)
+            he_img = Image.open(he_path)
+            ihc_path = os.path.join(ihc_dir, img)
+            if not os.path.isfile(ihc_path):
+                ihc_path = os.path.join(ihc_dir, img.replace('.png', '.jpg'))
+            ihc_img = Image.open(ihc_path)
+            if os.path.getsize(ihc_path) < 100 or os.path.getsize(he_path) < 100:
+                os.remove(ihc_path)
+                os.remove(he_path)
+            if os.path.getsize(ihc_path) < 800 or os.path.getsize(he_path) < 800:
+                he_img.save(ihc_path)
+            img1_gray = he_img.convert('L')
+            img2_gray = ihc_img.convert('L')
+            img1_gray_np = np.array(img1_gray)
+            img2_gray_np = np.array(img2_gray)
+            ssim_value = ssim(img1_gray_np, img2_gray_np)
+            logger.info(f'{i}, ssim {ssim_value}')
+            if is_solid_color_numpy(ihc_path):
+                os.remove(ihc_path)
+                os.remove(he_path)
+                print(f'remove {ihc_path}')
+                continue
+            elif ssim_value >= 0.6:
+                logger.info('skip {}'.format(img))
+                continue
+            elif ssim_value <= 0.6 and random.random() < 0.2:
+                os.remove(ihc_path)
+                os.remove(he_path)
+                print(f'remove {ihc_path}')
+                continue
+            elif ssim_value <= 0.5 and random.random() < 0.2:
+                os.remove(ihc_path)
+                os.remove(he_path)
+                print(f'remove {ihc_path}')
+                continue
+            elif ssim_value <= 0.4 and random.random() < 0.2:
+                os.remove(ihc_path)
+                os.remove(he_path)
+                print(f'remove {ihc_path}')
+                continue
+            elif ssim_value <= 0.3 and random.random() < 0.2:
+                os.remove(ihc_path)
+                os.remove(he_path)
+                print(f'remove {ihc_path}')
+                continue
+            elif ssim_value <= 0.2 and random.random() < 0.2:
+                os.remove(ihc_path)
+                os.remove(he_path)
+                print(f'remove {ihc_path}')
+                continue
+            elif ssim_value <= 0.1 and random.random() < 0.2:
+                os.remove(ihc_path)
+                os.remove(he_path)
+                print(f'remove {ihc_path}')
+                continue
+            elif ssim_value <= 0.05 and random.random() < 0.2:
+                os.remove(ihc_path)
+                os.remove(he_path)
+                print(f'remove {ihc_path}')
+                continue
+            plt.figure(figsize=(12, 6))  # 可以根据需要调整窗口大小
+
+            # 显示第一张图片
+            plt.subplot(1, 2, 1)  # 1行2列的第1个
+            plt.imshow(he_img)
+            plt.axis('off')  # 不显示坐标轴
+
+            # 显示第二张图片
+            plt.subplot(1, 2, 2)  # 1行2列的第2个
+            plt.imshow(ihc_img)
+            plt.axis('off')  # 不显示坐标轴
+
+            # 显示图形
+            plt.show()
+            i = input("input:")
+            if not i:
+                os.remove(ihc_path)
+                os.remove(he_path)
+                print(f'remove {ihc_path}')
+            gc.collect()
+        except:
+            traceback.print_exc()
+            pass
+
+
+# manual_check()
+# size_filter(1000)
+# split_data_a()
+def trank_check():
+    images = os.listdir(he_dir)
+    for i, img in enumerate(images):
+        try:
+            he_path = os.path.join(he_dir, img)
+            ihc_path = os.path.join(ihc_dir, img)
+            if not os.path.isfile(ihc_path):
+                ihc_path = os.path.join(ihc_dir, img.replace('.png', '.jpg'))
+            he_img = Image.open(he_path)
+            ihc_img = Image.open(ihc_path)
+        except:
+            os.remove(ihc_path)
+            os.remove(he_path)
+            print(f'remove{i} {ihc_path}')
+
+
+# trank_check()
+def test():
+    he_dir = rf'/NAS2/Data1/lbliao/Data/MSI/pair/4096/P2/trainA'
+    ihc_dir = rf'/NAS2/Data1/lbliao/Data/MSI/pair/4096/P2/trainB'
+    for A_img in os.listdir(he_dir):
+        A_path = os.path.join(he_dir, A_img)
+        B_path = os.path.join(ihc_dir, A_img)
+        try:
+            A_img = Image.open(A_path).convert('RGB')
+            B_img = Image.open(B_path).convert('RGB')
+        except Exception:
+            os.remove(A_path)
+            os.remove(B_path)
+
+logger.info('start')
+# test()
+exts = ['M1', 'M2', 'M6', 'P2']
+for ext in exts:
+    reg_dir = rf'/NAS2/Data1/lbliao/Data/MSI/0327/pair/4096/{ext}/dhr/reg_ihc'
+    files = os.listdir(reg_dir)
+    for file in files:
+        reg_path = os.path.join(reg_dir, file)
+        if os.path.isfile(reg_path):
+            os.rename(reg_path, reg_path.replace('.tiff', '.png'))
+        elif os.path.isdir(reg_path):
+            shutil.rmtree(reg_path)
+        logger.info(f'process {file}')
+    logger.info(f'process {ext}')
